refactor(download): rename to bufferSave
